+++ web/src/pages/home/component/index.js
@@ -7,9 +7,18 @@ import Spinner from '../../../components/spinner';
 
 export default props => <div>
     <Menu />
+    <div className="headerContainer">
+        <div className="headerTextContainer">
+            <h1>HEADER</h1>
+            <h6 className="thin">THIS IS A HEADER</h6>
+        </div>
+        <div className="canvasContainer">
+            <CanvasHeader />
+        </div>
+    </div>
+
     <div className="mainContainer">
         <h1>Hola Mundo!</h1>
-        <CanvasHeader />
         <p>{props.data}</p>
         <button type="button" className="btn btn-primary" onClick={() => props.sendMessage('mensaje nuevo')}>agregar dato</button><br />
         <Link href="/contact">
@@ -17,8 +26,8 @@ export default props => <div>
         </Link>
         <Spinner
             type={'spinner-grow'}
-            color = {'blue'}
-            align = {'text-center'}
+            color={'blue'}
+            align={'text-center'}
         />
     </div>
 
@@ -29,5 +38,31 @@ export default props => <div>
     }
     .mainContainer {
         padding: 20px;
+    }
+    .headerContainer {
+        background-image: url(/static/backgroundDemo.jpg);
+        position: relative;
+        width: 100%;
+        overflow: hidden;
+        background-size: cover;
+        background-position: center center;
+        z-index: 1;
+    }
+    .headerTextContainer {
+        position: absolute;
+        margin: 0;
+        padding: 0;
+        color: #ffffff;
+        text-align: center;
+        top: 50%;
+        left: 50%;
+        -webkit-transform: translate3d(-50%,-50%,0);
+        transform: translate3d(-50%,-50%,0);
+    }
+    .canvasContainer {
+        background-color: #33333329;
+    }
+    .thin {
+        font-weight: 200;
     }`}</style>
 </div>